django_diploma/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
diploma-frontend @ file:///C:/Users/BIRZHAN/PycharmProjects/python_django_diploma/diploma-frontend/dist/diploma-frontend-0.6.tar.gz#sha256=445b01fc4a7e391b2a0989aa23ba4daf09212bcc5b971f35729b80a8f74fecdb
Django==4.2.7
django-debug-toolbar==4.3.0
django-redis==5.4.0
django-request-logging==0.7.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
executing==2.0.1
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.18.0
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
